Extract AutoMergeController from IssueTracker (CROW-1094)
Move the auto-merge watcher (CROW-299) — applyAutoMerge, evaluateAutoMerge,
publishWatcherOffVerdict, the enable / direct-merge / update-branch attempts,
recordAutoMergeSuccess, reevaluateAutoMergeAfterLabel, ensureMergeLabel*, the
verdict/notify helpers, and the eligibility / skip-reason / gate statics plus
AutoMergeSkipReason / AutoMergeOutcome — into its own collaborator, owning all
its merge bookkeeping. The shared codeBackend / prHasCrowAuthoredCommit helpers
live here now and are re-exposed on the tracker (instance forwarders) for the
rebase / re-review watchers. Dispatch Tasks use [weak self].

applyPRStatuses and addMergeLabel drive it; an extension IssueTracker re-exposes
the IssueTracker.<symbol> / tracker.<member> surface the tests use (skip-reason
+ outcome types, the gate/eligibility statics, the merge-label constant, and the
in-flight / update-branch-attempted / failure-count proxies). extractCrowSessionUUIDs
and buildPRStatus stay on the tracker (shared with attribution / PR-status).

IssueTracker.swift is now a ~1.5k-line orchestrator (was 5,428).
swift test --package-path Packages/CrowEngine: 789 tests pass.


Crow-Session: 18A4CD55-8B33-48E4-B5EB-E710314DF5E5
